pyproject.toml
requires = [
    "setuptools"
]
build-backend = "setuptools.build_meta" 

[project]
name = "mqtt-topping-py"
version = "0.9.1"
description = "Syntactical sugar on top of the MQTT client cake"
authors = [
  {name = "Julian Krumow", email = "julian.krumow@artcom.de"},
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt"
]

[tool.pytest.ini_options]
minversion = "8.3"
addopts = "-ra"
testpaths = [
    "tests"
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
